+++ package.json
@@ -3,10 +3,19 @@
   "version": "0.6.0",
   "description": "MathJax renderer plugin for Hexo",
   "main": "index.js",
+  "scripts": {
+    "test": "mocha"
+  },
   "dependencies": {
     "ejs": "^1.0.0"
   },
-  "devDependencies": {},
+  "devDependencies": {
+    "chai": "^3.4.1",
+    "ejs": "^1.0.0",
+    "hexo": "^3.1.1",
+    "hexo-fs": "^0.1.5",
+    "mocha": "^2.3.4"
+  },
   "repository": {
     "type": "git",
     "url": "https://github.com/phoenixcw/hexo-renderer-mathjax.git"
